Handle randomly empty commit
Commits work by randomly generating a number of lines in a file, then
adding and committing the file. There's a chance with this process that
it will generate the same lines that were in the file before, and thus
not actually change anything. Rare enough that I think we can just
rescue and skip.